+++ src/index.js
@@ -42,7 +42,6 @@ export * from './components/ResizeSensor';
 export * from './components/Scrollbar'; // deprecated
 export * from './components/SectionPreloader';
 export * from './components/Select';
-export * from './components/SideMenu';
 export * from './components/SplashError';
 export * from './components/SplashModal';
 export * from './components/Spin';